backend/requirements.txt
# Django Core
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0

# Authentication
djangorestframework-simplejwt>=5.4,<6.0

# Environment
python-dotenv>=1.0.0

# Database (optional - using SQLite by default)
# psycopg2-binary>=2.9  # PostgreSQL
# mysqlclient>=2.2  # MySQL

# For production (optional)
# gunicorn>=23.0
# whitenoise>=6.8  # Static file serving